rename some classes, refactor ListCell
