+++ Vrooom/Model/OperatedCar.swift
@@ -11,6 +11,5 @@ import Foundation
 class OperatedCar {
     static var index: Int?
     static var newCar = false
-    static var imagePath: String? = nil
 }
